-split header into a little bit pices - edit home page
